requirements.txt
alembic==1.11.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
flasgger==0.9.7.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flask-swagger-ui==4.11.1
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
mistune==2.0.5
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
Werkzeug==2.3.4
